Changed macro to work with signed 64-bit values

+++ src/lib.rs
@@ -17,7 +17,7 @@
 /// #[macro_use]
 /// extern crate alt_bitflags;
 ///
-/// struct Something(u64);
+/// struct Something(i64);
 ///
 /// impl Something {
 ///     flag_ro!(present, 0);
@@ -67,7 +67,7 @@ macro_rules! flag_ro {
 /// #[macro_use]
 /// extern crate alt_bitflags;
 ///
-/// struct Something(u64);
+/// struct Something(i64);
 ///
 /// impl Something {
 ///     flag_rw!(present, set_present, 0);
@@ -92,7 +92,7 @@ macro_rules! flag_rw {
         #[inline="always"]
         pub fn $set_name(&mut self, b: bool) {
             let x = b as i64;
-            self.0 ^= ((-x ^ self.0 as i64) & (1 << $pos)) as u64;
+            self.0 ^= ((-x ^ self.0) & (1 << $pos));
         }
     );
